Add files via upload
This is machine learning approach to learn from a set of cell types and predict the cell types in the test dataset with unknown cell types. The method uses RandomForest algorithm in R. 
It can be used splitting a Seurat object into test and training or between 2 different Seurat object with at least some expected similar cell types.